+++ Runtime/EventManagerBase.cs
@@ -3,21 +3,20 @@ using System.Collections.Generic;
 
 namespace DJM.EventManager
 {
-    public abstract class EventManagerBase<T> where T : MulticastDelegate
+    public abstract class EventManagerBase<TEventId, TDelegate> : IEventManager<TEventId, TDelegate> where TDelegate : Delegate
     {
-        protected readonly IDictionary<int, T> EventTable;
+        protected readonly IDictionary<TEventId, TDelegate> EventTable;
 
         protected EventManagerBase()
         {
-            EventTable = new Dictionary<int, T>();
+            EventTable = new Dictionary<TEventId, TDelegate>();
         }
+        
+        public abstract void AddObserver(TEventId eventId, TDelegate handler);
 
+        public abstract void RemoveObserver(TEventId eventId, TDelegate handler);
 
-        public abstract void AddObserver(int eventId, T handler);
-
-        public abstract void RemoveObserver(int eventId, T handler);
-
-        public void ClearEvent(int eventId)
+        public void ClearEvent(TEventId eventId)
         {
             if( EventTable.ContainsKey(eventId)) 
                 EventTable.Remove(eventId);