+++ Runtime/EventMessageBase.cs
@@ -1,0 +1,9 @@
+namespace DJM.EventManager
+{
+    public struct EventMessageBase
+    {
+        public string Something;
+    }
+}
+
+